+++ R/print.fdt.quantile.R
@@ -1,0 +1,6 @@
+print.fdt.quantile <- function(x, ...)
+{
+  print(unclass(x), ...)
+
+  invisible(x)
+}