+++ R/print.fdt.quantile.multiple.R
@@ -1,0 +1,10 @@
+print.fdt.quantile.multiple <- function(x, ...)
+{
+  for (nm in names(x)) {
+    cat(nm, '\n')
+    print(x[[nm]], ...)
+    cat('\n')
+  }
+
+  invisible(x)
+}