+++ R/xtable.fdt.quantile.R
@@ -1,0 +1,18 @@
+xtable.fdt.quantile <- function(x,
+                                caption = NULL,
+                                label = NULL,
+                                align = NULL,
+                                digits = NULL,
+                                display = NULL,
+                                auto = FALSE, ...)
+{
+  res_DF <- .fdt.xtable.prep.quantile(x)
+
+  return(xtable(res_DF,
+                caption = caption,
+                label = label,
+                align = align,
+                digits = digits,
+                display = display,
+                auto = auto, ...))
+}